refactor(airlock): merge storage into coordinator, event bus architecture
Merge ActionStorage into ActionCoordinator so the coordinator owns the
full lifecycle: SQLite storage, pending decision futures, and a generic
event bus. Both MCP and REST faces subscribe to coordinator events and
translate them into their own notification mechanisms.

Key changes:
- Delete storage.py; move ORM models and CRUD into coordinator.py
- Add CoordinatorEvent types (ActionCreatedEvent, ActionUpdatedEvent)
  with async listener callback pattern
- MCP server subscribes to coordinator events for resource-updated
  notifications instead of driving storage directly
- SSE subscriber queues are local to operator_api.py, not on coordinator
- Coordinator.create_action/update_and_log emit events after mutations
- App.py owns coordinator lifecycle (init/close) wrapping MCP lifespan
- Fix ActionKey.__str__ for use in log messages
- Remove client_id/subject defaults from Action model
- Require body: RejectBody on reject endpoint (no default None)
